fix: resolve all clippy warnings across workspace
Fix collapsible_if, redundant closures, unnecessary casts, needless
borrows, clamp patterns, explicit counter loops, derivable impls, and
remove dead code to pass CI clippy checks.

+++ crates/debugger/tests/debugger.rs
@@ -176,7 +176,7 @@ async fn test_remote_attach() -> eyre::Result<()> {
                             ..
                         },
                         ..
-                    } if *path == file_path && line as usize == breakpoint_line
+                    } if *path == file_path && line == breakpoint_line
                 ));
                 break;
             }
@@ -292,7 +292,7 @@ async fn test_debugger() -> eyre::Result<()> {
                             ..
                         },
                         ..
-                    } if *path == file_path && line as usize == breakpoint_line
+                    } if *path == file_path && line == breakpoint_line
                 ));
                 break;
             }
